refactor: 도메인 로직으로 삽입

+++ src/main/java/com/snackgame/server/member/controller/dto/StatusResponse.java
@@ -16,6 +16,6 @@ public class StatusResponse {
     public static StatusResponse of(Status status) {
         return new StatusResponse(
                 status.getLevel(),
-                status.getExp());
+                status.getExp().doubleValue());
     }
 }
